fixed add new tech bug after merge.
